Fixed duplicate grid question mentioned in Issue 7
Was not correctly entering the starting display order.  Added
parameter to TextboxToRecord initializer.

This also fixes the issue of grid questions being placed in the wrong
spot.

Also noticed that delete was not cleaning up spare rows in the grid.
Question has a belongs_to dependent => :destroy on QuestionGroup, so if you delete
the question, the question group goes.  But the QuestionGroup does not
have a corresponding dependent => :destroy, so a deleted QuestionGroup may
have orphans.  Added cleanup to the QuestionController destroy method to
allow for this.

	modified:   app/controllers/questions_controller.rb
	modified:   lib/surveyor_gui/models/question_methods.rb

+++ app/controllers/questions_controller.rb
@@ -71,6 +71,11 @@ class QuestionsController < ApplicationController
       render :text=>"The following questions have logic that depend on this question: \n\n"+question.dependency_conditions.map{|d| " - "+d.dependency.question.text}.join('\n')+"\n\nPlease delete logic before deleting this question.".html_safe
       return
     end
+    if question.part_of_group?
+      question.question_group.questions.each{|q| q.destroy}
+      render :text=>''
+      return
+    end
     question.destroy
     render :text=>''
   end
